+++ legwork/snr.py
@@ -61,16 +61,16 @@ def snr_circ_stationary(m_c, f_orb, dist, t_obs, position=None, polarisation=Non
     # only need to compute n=2 harmonic for circular
     h_0_circ_2 = strain.h_0_n(m_c=m_c, f_orb=f_orb, ecc=np.zeros_like(f_orb).value, n=2, dist=dist,
                               position=position, polarisation=polarisation, inclination=inclination,
-                              interpolated_g=interpolated_g).flatten() ** 2
+                              interpolated_g=interpolated_g).flatten()**2
 
     h_f_src_circ_2 = h_0_circ_2 * t_obs
     if interpolated_sc is not None:
         h_f_lisa_2 = interpolated_sc(2 * f_orb)
     else:
         h_f_lisa_2 = psd.power_spectral_density(f=2 * f_orb, t_obs=t_obs, instrument=instrument,
-                                                custom_function=custom_psd,
+                                                custom_psd=custom_psd,
                                                 position=position, polarisation=polarisation)
-    snr = (h_f_src_circ_2 / h_f_lisa_2) ** 0.5
+    snr = (h_f_src_circ_2 / h_f_lisa_2)**0.5
 
     return snr.decompose()
 
@@ -149,7 +149,7 @@ def snr_ecc_stationary(m_c, f_orb, ecc, dist, t_obs, harmonics_required,
     # calculate source signal
     h_0_ecc_n_2 = strain.h_0_n(m_c=m_c, f_orb=f_orb, ecc=ecc, n=n_range, dist=dist,
                                position=position, polarisation=polarisation,
-                               inclination=inclination, interpolated_g=interpolated_g) ** 2
+                               inclination=inclination, interpolated_g=interpolated_g)**2
 
     # reshape the output since only one timestep
     h_0_ecc_n_2 = h_0_ecc_n_2.reshape(len(m_c), harmonics_required)
@@ -162,7 +162,7 @@ def snr_ecc_stationary(m_c, f_orb, ecc, dist, t_obs, harmonics_required,
         h_f_lisa_n_2 = h_f_lisa_n_2.reshape(f_n.shape)
     else:
         h_f_lisa_n_2 = psd.power_spectral_density(f=f_n, t_obs=t_obs,
-                                                  instrument=instrument, custom_function=custom_psd,
+                                                  instrument=instrument, custom_psd=custom_psd,
                                                   position=position, polarisation=polarisation)
 
     snr_n_2 = (h_f_src_ecc_2 / h_f_lisa_n_2).decompose()
@@ -171,7 +171,7 @@ def snr_ecc_stationary(m_c, f_orb, ecc, dist, t_obs, harmonics_required,
         return snr_n_2
 
     # calculate the signal-to-noise ratio
-    snr = (np.sum(snr_n_2, axis=1)) ** 0.5
+    snr = (np.sum(snr_n_2, axis=1))**0.5
 
     if ret_max_snr_harmonic:
         max_snr_harmonic = np.argmax(snr_n_2, axis=1) + 1
@@ -256,7 +256,7 @@ def snr_circ_evolving(m_1, m_2, f_orb_i, dist, t_obs, n_step,
 
     # calculate the characteristic power
     h_c_n_2 = strain.h_c_n(m_c=m_c, f_orb=f_orb_evol, ecc=np.zeros_like(f_orb_evol).value, n=2, dist=dist,
-                           interpolated_g=interpolated_g) ** 2
+                           interpolated_g=interpolated_g)**2
     h_c_n_2 = h_c_n_2.reshape(len(m_c), n_step)
 
     # calculate the characteristic noise power
@@ -265,11 +265,11 @@ def snr_circ_evolving(m_1, m_2, f_orb_i, dist, t_obs, n_step,
         h_f_lisa_2 = h_f_lisa_2.reshape(f_orb_evol.shape)
     else:
         h_f_lisa_2 = psd.power_spectral_density(f=2 * f_orb_evol, t_obs=t_obs,
-                                                instrument=instrument, custom_function=custom_psd,
+                                                instrument=instrument, custom_psd=custom_psd,
                                                 position=position, polarisation=polarisation)
-    h_c_lisa_2 = (2 * f_orb_evol) ** 2 * h_f_lisa_2
+    h_c_lisa_2 = (2 * f_orb_evol)**2 * h_f_lisa_2
 
-    snr = np.trapz(y=h_c_n_2 / h_c_lisa_2, x=2 * f_orb_evol, axis=1) ** 0.5
+    snr = np.trapz(y=h_c_n_2 / h_c_lisa_2, x=2 * f_orb_evol, axis=1)**0.5
 
     return snr.decompose()
 
@@ -370,7 +370,7 @@ def snr_ecc_evolving(m_1, m_2, f_orb_i, dist, ecc, harmonics_required, t_obs, n_
     t_evol = np.minimum(t_merge - t_before, t_obs).to(u.s)
     # get eccentricity and f_orb evolutions
     e_evol, f_orb_evol = evol.evol_ecc(ecc_i=ecc, t_evol=t_evol, n_step=n_step, m_1=m_1, m_2=m_2,
-                                       f_orb_i=f_orb_i, n_proc=n_proc, t_before=t_before)
+                                       f_orb_i=f_orb_i, n_proc=n_proc, t_before=t_before, t_merge=t_merge)
 
     maxes = np.where(np.logical_and(e_evol == 0.0, f_orb_evol == 1e2 * u.Hz),
                      -1 * u.Hz, f_orb_evol).max(axis=1)
@@ -384,17 +384,17 @@ def snr_ecc_evolving(m_1, m_2, f_orb_i, dist, ecc, harmonics_required, t_obs, n_
     # calculate the characteristic strain
     h_c_n_2 = strain.h_c_n(m_c=m_c, f_orb=f_orb_evol, ecc=e_evol, n=harms, dist=dist,
                            position=position, polarisation=polarisation, inclination=inclination,
-                           interpolated_g=interpolated_g) ** 2
+                           interpolated_g=interpolated_g)**2
 
     # calculate the characteristic noise power
     if interpolated_sc is not None:
         h_f_lisa = interpolated_sc(f_n_evol.flatten())
     else:
         h_f_lisa = psd.power_spectral_density(f=f_n_evol.flatten(), t_obs=t_obs,
-                                              instrument=instrument, custom_function=custom_psd,
+                                              instrument=instrument, custom_psd=custom_psd,
                                               position=position, polarisation=polarisation)
     h_f_lisa = h_f_lisa.reshape(f_n_evol.shape)
-    h_c_lisa_2 = f_n_evol ** 2 * h_f_lisa
+    h_c_lisa_2 = f_n_evol**2 * h_f_lisa
 
     snr_evol = h_c_n_2 / h_c_lisa_2